# chore: improve coverage for secret option