--- a/test/secret.js
+++ b/test/secret.js
@@ -2,6 +2,37 @@
 
 const tap = require('tap')
 
+tap.test('throws when secret is less than 32 bytes', async function (t) {
+  t.plan(2)
+
+  const fastify = require('fastify')({
+    logger: false
+  })
+  t.teardown(fastify.close.bind(fastify))
+
+  await fastify.register(require('../'), {
+    secret: 'a'.repeat(31)
+  }).after((err) => {
+    t.type(err, Error)
+    t.equal(err.message, 'secret must be at least 32 bytes')
+  })
+})
+
+tap.test('not throws when secret is greater than or equal to 32 bytes', async function (t) {
+  t.plan(1)
+
+  const fastify = require('fastify')({
+    logger: false
+  })
+  t.teardown(fastify.close.bind(fastify))
+
+  await fastify.register(require('../'), {
+    secret: 'a'.repeat(32)
+  }).after((err) => {
+    t.error(err)
+  })
+})
+
 tap.test('using secret without salt', function (t) {
   const fastify = require('fastify')({
     logger: false