This PR improves test coverage.

<!--
Thank you for your pull request. Please provide a description above and review
the requirements below.

Bug fixes and new features should include tests and possibly benchmarks.

Contributors guide: https://github.com/fastify/fastify/blob/master/CONTRIBUTING.md

By making a contribution to this project, I certify that:

* (a) The contribution was created in whole or in part by me and I
  have the right to submit it under the open source license
  indicated in the file; or

* (b) The contribution is based upon previous work that, to the best
  of my knowledge, is covered under an appropriate open source
  license and I have the right under that license to submit that
  work with modifications, whether created in whole or in part
  by me, under the same open source license (unless I am
  permitted to submit under a different license), as indicated
  in the file; or

* (c) The contribution was provided directly to me by some other
  person who certified (a), (b) or (c) and I have not modified
  it.

* (d) I understand and agree that this project and the contribution
  are public and that a record of the contribution (including all
  personal information I submit with it, including my sign-off) is
  maintained indefinitely and may be redistributed consistent with
  this project or the open source license(s) involved.
-->

#### Checklist

- [x] run `npm run test` and `npm run benchmark`
- [x] tests and/or benchmarks are included
- [ ] documentation is changed or added
- [x] commit message and code follows the [Developer's Certification of Origin](https://github.com/fastify/.github/blob/master/CONTRIBUTING.md#developers-certificate-of-origin-11)
      and the [Code of conduct](https://github.com/fastify/.github/blob/master/CODE_OF_CONDUCT.md)

before:

```
$ npm test

> @fastify/secure-session@5.1.0 test
> standard && tap test/*.js && tsd

 PASS  test/decorators.js 9 OK 66.974ms
 PASS  test/delete.js 14 OK 201.459ms
 PASS  test/cookie-name.js 6 OK 202.445ms
 PASS  test/dynamic-cookie-options.js 13 OK 225.886ms
 PASS  test/access.js 20 OK 240.085ms
 PASS  test/key.js 8 OK 230.739ms
 PASS  test/key-rotation.js 11 OK 250.853ms
 PASS  test/key-base64.js 5 OK 199.263ms
 PASS  test/malformed-cookie.js 5 OK 177.691ms
 PASS  test/nonce-bad-length.test.js 5 OK 141.767ms
 PASS  test/path.js 4 OK 151.018ms
 PASS  test/salt.js 5 OK 146.278ms
 PASS  test/schema.js 6 OK 171.06ms
 PASS  test/secret.js 15 OK 2s



  🌈 SUMMARY RESULTS 🌈


Suites:   14 passed, 14 of 14 completed
Asserts:  126 passed, of 126
Time:   4s
ERROR: Coverage for lines (92.74%) does not meet global threshold (100%)
ERROR: Coverage for branches (89.06%) does not meet global threshold (100%)
ERROR: Coverage for statements (92.8%) does not meet global threshold (100%)
----------|---------|----------|---------|---------|---------------------------------
File      | % Stmts | % Branch | % Funcs | % Lines | Uncovered Line #s
----------|---------|----------|---------|---------|---------------------------------
All files |    92.8 |    89.06 |     100 |   92.74 |
 index.js |    92.8 |    89.06 |     100 |   92.74 | 38,52,70,81,118-119,141-142,262
----------|---------|----------|---------|---------|---------------------------------
```

after:

```
$ npm test

> @fastify/secure-session@5.1.0 test
> standard && tap test/*.js && tsd

 PASS  test/decorators.js 9 OK 58.74ms
 PASS  test/dynamic-cookie-options.js 13 OK 232.953ms
 PASS  test/delete.js 14 OK 205.353ms
 PASS  test/key-rotation.js 11 OK 252.425ms
 PASS  test/key-base64.js 5 OK 214.142ms
 PASS  test/key.js 8 OK 245.443ms
 PASS  test/access.js 20 OK 262.877ms
 PASS  test/cookie-name.js 6 OK 215.743ms
 PASS  test/malformed-cookie.js 5 OK 194.318ms
 PASS  test/nonce-bad-length.test.js 5 OK 149.038ms
 PASS  test/path.js 4 OK 157.78ms
 PASS  test/salt.js 5 OK 146.809ms
 PASS  test/schema.js 6 OK 168.078ms
 PASS  test/secret.js 18 OK 3s



  🌈 SUMMARY RESULTS 🌈


Suites:   14 passed, 14 of 14 completed
Asserts:  129 passed, of 129
Time:   5s
ERROR: Coverage for lines (93.54%) does not meet global threshold (100%)
ERROR: Coverage for branches (90.62%) does not meet global threshold (100%)
ERROR: Coverage for statements (93.6%) does not meet global threshold (100%)
----------|---------|----------|---------|---------|------------------------------
File      | % Stmts | % Branch | % Funcs | % Lines | Uncovered Line #s
----------|---------|----------|---------|---------|------------------------------
All files |    93.6 |    90.62 |     100 |   93.54 |
 index.js |    93.6 |    90.62 |     100 |   93.54 | 52,70,81,118-119,141-142,262
----------|---------|----------|---------|---------|------------------------------
```